Get Games

get_current_games provides an API to return the current rounds matches in a data frame.

get_games(user, pass, comp, round = NULL)

Arguments

user
monash username, in text

pass
monash password, in text

comp
comp type, should be one of "normal", "gauss" or "info"

round
(optional), round number to return. If not provided, will try find the current round.

Value

data.frame of the matches from the currently available tipping round

Examples

if (FALSE) {
get_games(user, pass, comp)
}