Submit Tips

Submit Tips

submit_tips(games, user, pass, comp, round = NULL)

Arguments

games

a table of games, ideally returned from get_games

user

monash username, in text

pass

monash password, in text

comp

comp type, should be one of "normal", "gauss" or "info"

round

(optional), round number to return. If not provided, will try find the current round.

Value

A tibble, containing the games and tips submitted via the API with a status column

Examples

if (FALSE) {
submit_tips(games, user, pass, comp)
}